+++ packages/maker/flatpak/package.json
@@ -12,11 +12,13 @@
     "test:base": "cross-env TS_NODE_FILES=1 mocha --config ../../../.mocharc.js"
   },
   "devDependencies": {
+    "@electron-forge/test-utils": "7.6.0",
     "chai": "^4.3.3",
     "chai-as-promised": "^7.0.0",
     "mocha": "^9.0.1",
     "proxyquire": "^2.1.3",
-    "sinon": "^13.0.1"
+    "sinon": "^13.0.1",
+    "vitest": "^2.1.6"
   },
   "engines": {
     "node": ">= 16.4.0"
